requirements.txt
pydub
openai
google-cloud-texttospeech
fastapi
uvicorn[standard]
wheel
jsonschema
